ugtrain: introduce uintptr_t handling
Calulating with (void *) for memory addresses is really cumbersome
with a lot of conversions between pointers and integers of the size
of a pointer. In order to simplify that, store all memory addresses
as uintptr_t.

The printf() and scanf() format specifiers need to be changed to
PRIxPTR and SCNxPTR. Also the limits are different with
UINTPTR_MAX. But uintptr_t is optional and may not be available.
So use unsigned long in that case unless it is 64 bit Windows.
There we need unsigned long long.

For C++ there is also the integer to C++ string conversion which
has to handle the integer in HEX correctly. So introduce the
to_xstring() function. Also cout/cerr need to be changed as "0x"
and the hex conversion are missing.

+++ src/maps.h
@@ -60,7 +60,7 @@ static inline void list_regions (list<struct region> *rlist)
 	list_for_each (rlist, region)
 		printf("[%2u] %14lx, %7lu bytes, %5s, "
 			"%14lx, %c%c%c, %s\n", region->id,
-			(ulong) region->start, region->size,
+			region->start, region->size,
 			region_type_names[region->type], region->load_addr,
 			region->flags.read ? 'r' : '-',
 			region->flags.write ? 'w' : '-',
